% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{build_file_db_from_raw_tables}
\alias{build_file_db_from_raw_tables}
\title{Extract file metadata from the full set of ENCODE metadata tables.}
\usage{
build_file_db_from_raw_tables(tables)
}
\arguments{
\item{tables}{A list of ENCODE metadata tables as loaded by
fetch_and_clean_raw_ENCODE_tables.}
}
\value{
A \code{list} containing two \code{data.table} objects:
  \code{encode_df}, containing the most interesting metadata columns,
  and \code{encode_df_ext}, containing relevant metadata for all
  ENCODE files.
}
\description{
Extract file metadata from the full set of ENCODE metadata tables.
}
\examples{
    \dontrun{
        tables = fetch_and_clean_raw_ENCODE_tables()
        build_file_db_from_raw_tables(tables = tables)
    }
}
