% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api.R
\name{fetch_table_from_ENCODE_REST}
\alias{fetch_table_from_ENCODE_REST}
\title{Extract a data.frame corresponding to a table in ENCODE database}
\usage{
fetch_table_from_ENCODE_REST(type)
}
\arguments{
\item{type}{The type of table to extract from the ENCODE rest api.
Available types can be obtained using 
\code{\link{get_encode_types}}.}
}
\value{
a \code{data.frame} corresponding to the table asked. If no match is
        found, returns an empty \code{data.frame}
}
\description{
Extract a data.frame corresponding to a table in ENCODE database
}
\keyword{internal}
