% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{generate_encode_df_full}
\alias{generate_encode_df_full}
\title{Given the raw ENCODE tables, this generate a data.table with the
full set of file metadata columns.}
\usage{
generate_encode_df_full(tables)
}
\arguments{
\item{tables}{A list of ENCODE metadata tables as loaded by
fetch_and_clean_raw_ENCODE_tables.}
}
\value{
a \code{data.table} containing relevant metadata for all
  ENCODE files.
}
\description{
Given the raw ENCODE tables, this generate a data.table with the
full set of file metadata columns.
}
\examples{
    \dontrun{
        tables = fetch_and_clean_raw_ENCODE_tables()
        export_ENCODEdb_matrix(tables = tables)
    }

}
