% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{generate_encode_df_lite}
\alias{generate_encode_df_lite}
\title{Extract file metadata from the full set of ENCODE metadata tables.}
\usage{
generate_encode_df_lite(tables)
}
\arguments{
\item{tables}{A list of ENCODE metadata tables as loaded by
fetch_and_clean_raw_ENCODE_tables.}
}
\value{
a \code{data.table} containing relevant metadata for all
  ENCODE files.
}
\description{
Extract file metadata from the full set of ENCODE metadata tables.
}
\examples{
    \dontrun{
        tables = fetch_and_clean_raw_ENCODE_tables()
        export_ENCODEdb_matrix(tables = tables)
    }

}
