% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api.R
\name{get_schema_urls}
\alias{get_schema_urls}
\title{Returns the URLs for downloading the XML schemas from ENCODE's github.}
\usage{
get_schema_urls()
}
\value{
a \code{character} \code{vector} of schema download URLs.
}
\description{
Returns the URLs for downloading the XML schemas from ENCODE's github.
}
\examples{
  ENCODExplorerData:::get_schema_urls()
}
