% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODExplorerData.R
\name{encode_df_full}
\alias{encode_df_full}
\title{Extended metadata for the files made available by ENCODE database as a 
\code{\link{data.table}} object. See \code{inst/scripts/make-data.R}
for the generation process.}
\format{A data table}
\description{
Extended metadata for the files made available by ENCODE database as a 
\code{\link{data.table}} object. See \code{inst/scripts/make-data.R}
for the generation process.
}
\examples{
    library(AnnotationHub)
    hub <- AnnotationHub()
    myfiles <- query(hub, c("ENCODExplorerData", "Full"))
    myfiles[[1]]
}
