% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODExplorerData.R
\name{encode_df_lite}
\alias{encode_df_lite}
\title{Metadata for the files made available by ENCODE database as a 
\code{\link{data.table}} object. See \code{inst/scripts/make-data.R}
for the generation process.
The tables were generated with the \code{fetch_and_clean_raw_ENCODE_tables} function.}
\format{A data table}
\description{
Metadata for the files made available by ENCODE database as a 
\code{\link{data.table}} object. See \code{inst/scripts/make-data.R}
for the generation process.
The tables were generated with the \code{fetch_and_clean_raw_ENCODE_tables} function.
}
\examples{
  \dontrun{
    library(AnnotationHub)
    hub <- AnnotationHub()
    myfiles <- query(hub, "ENCODE File Metadata (Light, 2019-04-12 build)")
    myfiles[[1]]
  }
}
\seealso{
\code{\link{get_encode_types}} to get a list of possible types. Note
  that some of the types are empty tables that are not included in the
  database created with \code{\link{fetch_and_clean_raw_ENCODE_tables}} function.
}
